Subdivided LH interfaces + (stub) PB client for high-level version
