+++ microArch/driver/liquidhandling/compositerobotinstruction.go
@@ -1141,7 +1141,7 @@ func (ins *AspirateInstruction) Generate(ctx context.Context, policy *wtype.LHPo
 	return nil, nil
 }
 
-func (ins *AspirateInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *AspirateInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	volumes := make([]float64, len(ins.Volume))
 	for i, vol := range ins.Volume {
 		volumes[i] = vol.ConvertTo(wunit.ParsePrefixedUnit("ul"))
@@ -1209,7 +1209,7 @@ func (ins *DispenseInstruction) Generate(ctx context.Context, policy *wtype.LHPo
 	return nil, nil
 }
 
-func (ins *DispenseInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *DispenseInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	volumes := make([]float64, len(ins.Volume))
 	for i, vol := range ins.Volume {
 		volumes[i] = vol.ConvertTo(wunit.ParsePrefixedUnit("ul"))
@@ -1273,7 +1273,7 @@ func (ins *BlowoutInstruction) Generate(ctx context.Context, policy *wtype.LHPol
 	return nil, nil
 }
 
-func (ins *BlowoutInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *BlowoutInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	volumes := make([]float64, len(ins.Volume))
 	for i, vol := range ins.Volume {
 		volumes[i] = vol.ConvertTo(wunit.ParsePrefixedUnit("ul"))
@@ -1322,7 +1322,7 @@ func (ins *PTZInstruction) Generate(ctx context.Context, policy *wtype.LHPolicyR
 	return nil, nil
 }
 
-func (ins *PTZInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *PTZInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.ResetPistons(ins.Head, ins.Channel)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -1395,7 +1395,7 @@ func (ins *MoveInstruction) Generate(ctx context.Context, policy *wtype.LHPolicy
 	return nil, nil
 }
 
-func (ins *MoveInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *MoveInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.Move(ins.Pos, ins.Well, ins.Reference, ins.OffsetX, ins.OffsetY, ins.OffsetZ, ins.Plt, ins.Head)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -1480,7 +1480,7 @@ func (ins *MoveRawInstruction) Generate(ctx context.Context, policy *wtype.LHPol
 	return nil, nil
 }
 
-func (ins *MoveRawInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *MoveRawInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	logger.Fatal("Not yet implemented")
 	panic("Not yet implemented")
 }
@@ -1542,7 +1542,7 @@ func (ins *LoadTipsInstruction) Generate(ctx context.Context, policy *wtype.LHPo
 	return nil, nil
 }
 
-func (ins *LoadTipsInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *LoadTipsInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.LoadTips(ins.Channels, ins.Head, ins.Multi, ins.HolderType, ins.Pos, ins.Well)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -1607,7 +1607,7 @@ func (ins *UnloadTipsInstruction) Generate(ctx context.Context, policy *wtype.LH
 	return nil, nil
 }
 
-func (ins *UnloadTipsInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *UnloadTipsInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.UnloadTips(ins.Channels, ins.Head, ins.Multi, ins.HolderType, ins.Pos, ins.Well)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2418,7 +2418,7 @@ func (ins *SetPipetteSpeedInstruction) Generate(ctx context.Context, policy *wty
 	return nil, nil
 }
 
-func (ins *SetPipetteSpeedInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *SetPipetteSpeedInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.SetPipetteSpeed(ins.Head, ins.Channel, ins.Speed)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2460,7 +2460,7 @@ func (ins *SetDriveSpeedInstruction) Generate(ctx context.Context, policy *wtype
 	return nil, nil
 }
 
-func (ins *SetDriveSpeedInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *SetDriveSpeedInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.SetDriveSpeed(ins.Drive, ins.Speed)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2492,7 +2492,7 @@ func (ins *InitializeInstruction) Generate(ctx context.Context, policy *wtype.LH
 	return nil, nil
 }
 
-func (ins *InitializeInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *InitializeInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.Initialize()
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2524,7 +2524,7 @@ func (ins *FinalizeInstruction) Generate(ctx context.Context, policy *wtype.LHPo
 	return nil, nil
 }
 
-func (ins *FinalizeInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *FinalizeInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.Finalize()
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2563,7 +2563,7 @@ func (ins *WaitInstruction) Generate(ctx context.Context, policy *wtype.LHPolicy
 	return nil, nil
 }
 
-func (ins *WaitInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *WaitInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	ret := driver.Wait(ins.Time)
 	if !ret.OK {
 		return fmt.Errorf(" %d : %s", ret.Errorcode, ret.Msg)
@@ -2612,7 +2612,7 @@ func (ins *LightsOnInstruction) Generate(ctx context.Context, policy *wtype.LHPo
 	return nil, nil
 }
 
-func (ins *LightsOnInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *LightsOnInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: LightsOnInstruction")
 }
 
@@ -2655,7 +2655,7 @@ func (ins *LightsOffInstruction) Generate(ctx context.Context, policy *wtype.LHP
 	return nil, nil
 }
 
-func (ins *LightsOffInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *LightsOffInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: LightsOffInstruction")
 }
 
@@ -2698,7 +2698,7 @@ func (ins *OpenInstruction) Generate(ctx context.Context, policy *wtype.LHPolicy
 	return nil, nil
 }
 
-func (ins *OpenInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *OpenInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: OpenInstruction")
 }
 
@@ -2741,7 +2741,7 @@ func (ins *CloseInstruction) Generate(ctx context.Context, policy *wtype.LHPolic
 	return nil, nil
 }
 
-func (ins *CloseInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *CloseInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: CloseInstruction")
 }
 
@@ -2784,7 +2784,7 @@ func (ins *LoadAdaptorInstruction) Generate(ctx context.Context, policy *wtype.L
 	return nil, nil
 }
 
-func (ins *LoadAdaptorInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *LoadAdaptorInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: LoadAdaptor")
 }
 
@@ -2827,7 +2827,7 @@ func (ins *UnloadAdaptorInstruction) Generate(ctx context.Context, policy *wtype
 	return nil, nil
 }
 
-func (ins *UnloadAdaptorInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (ins *UnloadAdaptorInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	return fmt.Errorf(" %d : %s", anthadriver.NIM, "Not yet implemented: UnloadAdaptor")
 }
 
@@ -3143,7 +3143,7 @@ func (ins *MixInstruction) GetParameter(name string) interface{} {
 
 }
 
-func (mi *MixInstruction) OutputTo(driver LiquidhandlingDriver) error {
+func (mi *MixInstruction) OutputTo(driver LowLevelLiquidhandlingDriver) error {
 	vols := make([]float64, len(mi.Volume))
 
 	for i := 0; i < len(mi.Volume); i++ {
